fixed venue img input too
